See Detailed Description
Made some changes to better stabilize where dmrlink.py looks for the
csv files… not perfect, but better. Expect more changes.

Have waffled back and forth on how to handle peers we lose contact
with… de-reg for sure, but ignore them, or try forever (until we get a
peer list without them from the master) to re-register with the peer?
Settled on trying forever, but will add code to request a new peer-list
every few hours.